fix(ux): swap leave and end meeting trigger placements to prevent accidental room termination
The red footer button, adjacent to frequently used controls, acted as
end-meeting-for-everyone for moderators, while the per-user leave action
was in the header overlay menu. Swap the actions, labels and
confirmations (visuals unchanged): the footer button now always leaves
the meeting, and moderators end the room from the header menu instead.
