+++ src/components/header/menus.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { MenuItem, MenuItems } from '@headlessui/react';
 import { useTranslation } from 'react-i18next';
 
-import { useAppDispatch } from '../../store';
+import { store, useAppDispatch } from '../../store';
 import {
   updateShowKeyboardShortcutsModal,
   updateShowRoomSettingsModal,
@@ -15,6 +15,9 @@ interface IHeaderMenusProps {
 const HeaderMenus = ({ onOpenAlert }: IHeaderMenusProps) => {
   const dispatch = useAppDispatch();
   const { t } = useTranslation();
+  const isAdmin = useMemo(() => {
+    return !!store.getState().session.currentUser?.metadata?.isAdmin;
+  }, []);
 
   return (
     <MenuItems
@@ -49,10 +52,10 @@ const HeaderMenus = ({ onOpenAlert }: IHeaderMenusProps) => {
         <button
           type="button"
           className="h-9 md:h-10 w-full cursor-pointer flex items-center hover:bg-Gray-50 dark:hover:bg-dark-secondary2 data-[focus]:bg-Gray-50 dark:data-[focus]:bg-dark-secondary2 text-sm gap-2 leading-none font-medium text-Gray-950 dark:text-white px-2 md:px-3 rounded-lg transition-all duration-300 relative"
-          onClick={() => onOpenAlert('logout')}
+          onClick={() => onOpenAlert(isAdmin ? 'end' : 'logout')}
         >
           <i className="pnm-logout text-primary-color dark:text-Blue2-500 text-lg me-2 transition ease-in" />
-          {t('header.menus.logout')}
+          {isAdmin ? t('header.menus.end') : t('header.menus.logout')}
         </button>
       </MenuItem>
     </MenuItems>
